Web: reduce number of requests in `paginated-fetch`
Unfortunately, the nu language doesn't give us a variant of
`take while` that allows us to pass along the last page of
results. This commit gets the first page and returns it if there
is no second page. Otherwise, the original logic is used. This
significantly reduces the number of requests when only one page is
available for a paginated resource. This is often the case for
sections, assignments, and some other resources.